updated fitur template

+++ database/migrations/2026_02_06_105714_remove_canonical_url_from_posts_and_pages_tables.php
@@ -9,11 +9,15 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('posts', function (Blueprint $table) {
-            $table->dropColumn('canonical_url');
+            if (Schema::hasColumn('posts', 'canonical_url')) {
+                $table->dropColumn('canonical_url');
+            }
         });
 
         Schema::table('pages', function (Blueprint $table) {
-            $table->dropColumn('canonical_url');
+            if (Schema::hasColumn('pages', 'canonical_url')) {
+                $table->dropColumn('canonical_url');
+            }
         });
     }
